minimize Dockerfile
By regrouping all steps in a single one, we remove an intermediate
tarball, and therefore build size.

We also use `apt-get clean` to remove the fetched packages and clear
out the result of the `apt-get update`, which shaves off a few more
bytes.

Finally, we use the "buster-slim" image and avoid installing
"recommends".

 * Before: 2.29GB
 * After: 1.41GB