factor out hardcoded /tmp in scripts
The idea behind this is to eventually make that configurable on the
commandline or at least easier to change.

We use the variable names "dangerzone" and "safezone" because
that *is* what they eventually become through the wrapper script, and
it makes the code more obvious.

Eventually, we hope to be able to not require the intermediate script,
but for now this will have to do.