create user with a higher UID
With the current UID scheme, it's likely that the user created inside
the container has the same UID as the user on the host (ie. UID
1000).

This makes isolation less effective: if someone manages to escape the
container somehow, they will be able to run code as the main user on
the machine, which probably has all sorts of juicy things in their
$HOME directory (e.g. code execution through .bashrc, private keys
disclosure in .ssh, etc).

By picking a higher UID range, we really run the container as a
completely different user.