remove sudo, and run the entire container as the right user
Before this change, the container would run as root, and then the
wrapper scripts would use `sudo` to drop privileges. Considering there
has been, in the past, vulnerabilities in `sudo`, this seems less than
ideal. It also makes audits much more complicated because this
Dockerfile has a red flag: no "USER" directive.

Obviously, this breaks dangerzone itself, because the directories
shared by the frontend will not have the right permission and the
container won't be able to write files there. There are instructions
in the Dockerfile on how to workaround those issues, but those are far
from ideal: they basically require root on the host to create a
directory owned by the funky user.

An alternative would be to assume a single-user system (ie. that the
calling user is UID=1000), but then that provides lesser privilege
separation, because the container actually runs as a privileged user,
which likely has sudo access, juicy stuff in their ~/.ssh/ or
~/.gnupg/ directories, or at the very least have a writable ~/.bashrc
to drop backdoors into.